+++ test/CrossDBActivityLinkSpec.hs
@@ -69,6 +69,7 @@ mkActivity name loc exs =
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = loc
+        , activityLocationSource = LocationDeclared
         , activityUnit = "kg"
         , exchanges = exs
         , activityParams = M.empty